smooth interpolation for a lot of objects including entities, effects (.efx), shaders; increased amount of possible effects (increased possible amount of refEnts); fixed and cleaned sound code a bit; random seed is same for all cgame, renderer and engine and based on cgame demo time; and some other changes

+++ codemp/qcommon/q_shared.h
@@ -876,7 +876,7 @@ typedef struct
 	saberTrail_t	trail;
 	int			hitWallDebounceTime;
 	int			storageTime;
-	int			extendDebounce;
+	float		extendDebounce;
 } bladeInfo_t;
 #define MAX_BLADES 8
 
@@ -2950,7 +2950,7 @@ typedef struct SSkinGoreData_s
 {
 	vec3_t			angles;
 	vec3_t			position;
-	int				currentTime;
+	float			currentTime;
 	int				entNum;
 	vec3_t			rayDirection;	// in world space
 	vec3_t			hitLocation;	// in world space
