+++ codemp/rd-vanilla/tr_shader.cpp
@@ -2984,10 +2984,10 @@ static void FixRenderCommandList( int newShader ) {
 					int i;
 					drawSurf_t	*drawSurf;
 					shader_t	*shader;
-					int			fogNum;
-					int			entityNum;
-					int			dlightMap;
-					int			sortedIndex;
+					int64_t		fogNum;
+					int64_t		entityNum;
+					int64_t		dlightMap;
+					int64_t		sortedIndex;
 					const drawSurfsCommand_t *ds_cmd =  (const drawSurfsCommand_t *)curCmd;
 
 					for( i = 0, drawSurf = ds_cmd->drawSurfs; i < ds_cmd->numDrawSurfs; i++, drawSurf++ ) {
